+++ src/Main.svelte
@@ -4,13 +4,14 @@
     import EndpointsTable from "./Components/EndpointsTable.svelte";
     import ModalEndpoint from "./Components/ModalEndpoint.svelte";
     import ModalSettings from "./Components/ModalSettings.svelte";
+    import ModalHistory from "./Components/ModalHistory.svelte";
     import {loadEndpoints, deleteAll} from './Services/EndpointsService.js';
     import {loadSettings} from './Services/SettingsService.js';
-    import {StoreConstants} from './Types/Constants.js'
     //----- </Internal> -----//
 
     let modalEndpoint;
     let modalSettings;
+    let modalHistory;
 
     onMount(async () => {
         await loadSettings();
@@ -23,6 +24,10 @@
 
     function uiOnSettingsClicked(event) {
         modalSettings.showModal(event.detail);
+    }
+
+    function uiOnHistoryClicked(event) {
+        modalHistory.showModal(event.detail);
     }
 
     function uiOnDeleteAllClicked() {
@@ -55,9 +60,11 @@
     <EndpointsTable 
         on:addClicked={uiOnAddClicked}
         on:settingsClicked={uiOnSettingsClicked}
-        on:deleteAllClicked={uiOnDeleteAllClicked}/>
+        on:deleteAllClicked={uiOnDeleteAllClicked}
+        on:historyClicked={uiOnHistoryClicked}/>
     <ModalEndpoint bind:this={modalEndpoint}/>
     <ModalSettings bind:this={modalSettings}/>
+    <ModalHistory bind:this={modalHistory}/>
 
 </div>