+++ src/Types/EndpointStatus.js
@@ -1,0 +1,8 @@
+export default class EndpointStatus {
+    status,//Status from Types/Constants
+    lastChecked,//date
+    //errorDetails
+    statusCode,//number
+    description,//string
+    history//[]
+}